Fix: prevent NullPointerException (java:S4449 l.81 src/main/java/com/djamware/mynotes/controllers/NotesController.java)

+++ src/main/java/com/djamware/mynotes/controllers/NotesController.java
@@ -79,7 +79,9 @@ public class NotesController {
 	@RequestMapping("/notes/delete")
 	public String delete(@RequestParam Long id) {
 		Notes note = noteRepository.findById(id).orElse(null);
+		if(note != null){
 		noteRepository.delete(note);
+		}
 
 		return "redirect:/notes";
 	}
